+++ src/Paket.Unity3D/Program.fs
@@ -56,7 +56,7 @@ try
             let deps = Dependencies.Locate().DependenciesFile |> DependenciesFile.ReadFromFile
             let lock = deps.FindLockfile().FullName |> LockFile.LoadFrom
             let sources = deps.GetAllPackageSources()
-            Paket.Unity3D.InstallProcess.Install(sources,false,true,false,lock)
+            Paket.Unity3D.InstallProcess.Install(sources,InstallerOptions.Default,lock)
         | Command(Add, rest) -> 
             let results = commandArgs<AddArgs> rest
             let nuget = results.GetResult <@ AddArgs.Nuget @>
@@ -74,7 +74,8 @@ try
 with
 | exn when not (exn :? System.NullReferenceException) -> 
     Environment.ExitCode <- 1
-    traceErrorfn "Paket.Unity3D failed with:%s\t%s" Environment.NewLine exn.Message
+    traceErrorfn "Paket.Unity3D failed with:%s
+    t%s" Environment.NewLine exn.Message
 
     if verbose then
         traceErrorfn "StackTrace:%s  %s" Environment.NewLine exn.StackTrace